cabi-testgen: more enhancements to addr-taken params/returns
This commit introduces support for passing the address of params
and returns to another (no-inline) function, so as to insure that
the back end has to properly materialize its address. Also cut down
slightly on the fraction of addr-taken params: make an initial decision
as to whether any params or returns are address-taken, then decide
which ones if so.