cabi-testgen: separate reflect and regular call paths
Rather than doing both regular-call and reflect-call testing in a
single invocation of <package>.Caller(), do two calls and have each
call do a single flavor of testing. The benefit is that we won't have
the args to the regular call held live across the call due to the
post-dominating reflect call.