cabi-testgen: more values live across function calls
To try to catch problems with things live across calls, arrange for
the generator to randomly pick values that would ordinarily be emitted
as literals (e.g. int64(7) or "foo") and instead create a call that
returns that value. This means more function calls appearing in things
like map key slots or elements of composite literals.